requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
djongo==1.3.6
idna==3.3
psycopg2==2.9.1
pymongo==3.12.1
pytz==2021.3
requests==2.26.0
sqlparse==0.2.4
typing-extensions==3.10.0.2
urllib3==1.26.7
